w3c-reports/12092014/efl/runned-tests/cors/response-headers-expected.txt

CONSOLE MESSAGE: Refused to get unsafe header " x-custom-header"
CONSOLE MESSAGE: Refused to get unsafe header "Server"
CONSOLE MESSAGE: Refused to get unsafe header "X-Powered-By"
Response headers


PASS getResponseHeader: Expose Access-Control-Expose-Headers (x-custom-header-comma) 
PASS getResponseHeader: Expose second Access-Control-Expose-Headers (x-second-expose) 
PASS getResponseHeader: Don't trim whitespace 
FAIL getResponseHeader: x-custom-header bytes assert_equals: x-custom-header-bytes expected "â¦" but got "…"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/cors/response-headers.htm:27:26
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
check_response_header@http://localhost:8800/cors/response-headers.htm:19:9
global code@http://localhost:8800/cors/response-headers.htm:34:22)
PASS getResponseHeader: Exposed server field readable (Date) 
PASS getResponseHeader: Cache-Control: readable by default 
PASS getResponseHeader: Content-Language: readable by default 
PASS getResponseHeader: Expires: readable by default 
PASS getResponseHeader: Last-Modified: readable by default 
PASS getResponseHeader: Pragma: readable by default 
PASS getResponseHeader: Server: unreadable by default 
PASS getResponseHeader: X-Powered-By: unreadable by default 
FAIL getResponseHeader: Combined testing of cors response headers assert_equals: x-custom-header expected "test" but got "test, test"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/cors/response-headers.htm:70:26
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
FAIL getResponse: don't expose x-nonexposed assert_equals: x-custom-header expected "test" but got "test, test"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/cors/response-headers.htm:89:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/cors/response-headers.htm:85:5)
PASS getAllResponseHeaders: don't expose x-nonexposed 
